requirements.txt
streamlit
Pillow
opencv-contrib-python
mysql-connector-python
face-recognition
numpy == 1.23.5
cmake == 3.18
streamlit-option-menu
python -m pip install dlib-19.24.1-cp311-cp311-win_amd64.whl
pandas

# if you are facing issue related dlib just run last  command make sure you have a dlib file





